Create RaceConditionProblem.c
This code here produces the race condition problem.
The scenario used here describes the problem when there is only one runway but there are two planes in the same time trying to land there. The race condition problem now allows two planes to land at the same time in the same runway producing a collision between the two planes respectfully between two processes.
This of course needs a solution, for which we are going to use a semaphore to avoid race condition in the file named: "RCSemaphore.c".